chore: clean up wireguard-tunnel-provider target
